requirements.txt
pyttsx3==2.90
vosk==0.3.45
sounddevice==0.4.6
requests==2.31.0
python-dotenv==1.0.0